chore: remove vehicle snap shot synchronize with vehicle profile and update seat template for high functionality coding

+++ src/main/java/vn/com/routex/merchant/platform/domain/booking/model/Booking.java
@@ -20,7 +20,7 @@ public class Booking {
     private String bookingCode;
     private String routeId;
     private String customerId;
-    private Integer seatCount;
+    private Long seatCount;
     private BigDecimal totalAmount;
     private String merchantId;
     private String currency;
